somehow workingImprove language app's functionality for learning exercises
Changes are made to improve functionality in language learning application. The `Models/Assignment.cs` and `ApplicationContext.cs` files are updated to refine the model of assignment class and database context. The `Controllers/LanguageController.cs` is refactored to adopt some best practices like avoiding field names which starts with underscore (_). New repositories are also added for handling audios in the application. Additionally, some helper functions are added in existing repositories to enhance their functionality. Changes are necessary for overall improvement in the application's functionality.